% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_poplevel_frequency}
\alias{assign_poplevel_frequency}
\title{Function that assigns a category ('Rare','Common' etc) to population-specific germline frequencies}
\usage{
assign_poplevel_frequency(var_df, dbquery = "1KG", pop = "european",
  result_tag = "FREQ_EXAC_EUROPEAN")
}
\arguments{
\item{var_df}{data frame with variants}

\item{dbquery}{1KG or ExAC}

\item{pop}{population}

\item{result_tag}{name of result column}
}
\value{
var_df
}
\description{
Function that assigns a category ('Rare','Common' etc) to population-specific germline frequencies
}
