% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cna_segment_annotation}
\alias{cna_segment_annotation}
\title{Function that annotates CNV segment files (FACETS)}
\usage{
cna_segment_annotation(cna_file, logR_threshold_amplification,
  logR_threshold_homozygous_deletion, format = "tcga_legacy")
}
\arguments{
\item{cna_file}{CNV file name}

\item{format}{CNV call format}
}
\value{
cna_data
}
\description{
Function that annotates CNV segment files (FACETS)
}
