% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cnv_segment_annotation}
\alias{cnv_segment_annotation}
\title{Function that annotates CNV segment files (FACETS)}
\usage{
cnv_segment_annotation(cnv_file, logR_threshold_amplification,
  logR_threshold_homozygous_deletion, format = "tcga_legacy")
}
\arguments{
\item{cnv_file}{CNV file name}

\item{format}{CNV call format}
}
\value{
cnv_data
}
\description{
Function that annotates CNV segment files (FACETS)
}
