% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_tier_tsv}
\alias{generate_tier_tsv}
\title{Function that generates dense and tiered annotated variant datasets}
\usage{
generate_tier_tsv(tier1_variants, tier2_variants, tier3_variants,
  tier4_variants, tier5_variants, sample_name = "test")
}
\arguments{
\item{tier1_variants}{df with tier 1 variants}

\item{tier2_variants}{df with tier 2 variants}

\item{tier3_variants}{df with tier 3 variants}

\item{tier4_variants}{df with tier 4 variants}

\item{tier5_variants}{df with tier 5 variants}

\item{sample_name}{Sample identifier}
}
\value{
tsv_variants data frame with tier-annotated list of variants for tab-separated output
}
\description{
Function that generates dense and tiered annotated variant datasets
}
