% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_calls}
\alias{get_calls}
\title{Function that reads a VCF from pcgr pipeline}
\usage{
get_calls(vcf_gz_file, sample_id = NULL)
}
\arguments{
\item{vcf_gz_file}{}
}
\value{
vcf_data_df
}
\description{
Function that reads a VCF from pcgr pipeline
}
