% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_clinical_associations_civic_cbmdb}
\alias{get_clinical_associations_civic_cbmdb}
\title{Function that appends clinical annotations for somatic cancer variants}
\usage{
get_clinical_associations_civic_cbmdb(vcf_data_df, mapping = "exact",
  variant_origin = "Somatic Mutation")
}
\arguments{
\item{vcf_data_df}{data frame with variants}

\item{mapping}{- one of 'exact' (allele-specific) or 'approximate' (codon or exon-level biomarkers)}

\item{ncgc}{- logical indicating whether NCGC-specific tags are to be appended}
}
\value{
vcf_data_df
}
\description{
Function that appends clinical annotations for somatic cancer variants
}
